chore: refactor docs and remove unused files
