+++ examples/my-app/src/pages/index.tsx
@@ -19,7 +19,6 @@ export default function TypescriptSdkDemo() {
         password: "password",
       },
     });
-
     setToken(userToken);
   };
 
@@ -30,11 +29,11 @@ export default function TypescriptSdkDemo() {
   useEffect(() => {
     if (token) {
       client.Auth.getUserQuery()
-        .then((data) => {
+        .then((data: any) => {
           console.log("data", data);
           setUser(data);
         })
-        .catch((error) => {
+        .catch((error: any) => {
           console.log("error", error);
         });
 
@@ -42,11 +41,11 @@ export default function TypescriptSdkDemo() {
         pageSize: null,
         nextPageToken: null,
       })
-        .then((data) => {
+        .then((data: any) => {
           console.log("data", data);
           setPipelines(data);
         })
-        .catch((error) => {
+        .catch((error: any) => {
           console.log("error", error);
         });
     }
@@ -66,3 +65,60 @@ export default function TypescriptSdkDemo() {
     </>
   );
 }
+
+// Console Token Implementation
+
+// import { useEffect, useState } from "react";
+// import InstillClient, {
+//   Nullable,
+//   Pipeline,
+//   User,
+// } from "@instill-ai/typescript-sdk";
+
+// export default function TypescriptSdkDemo() {
+//   const [pipelines, setPipelines] = useState<Pipeline[]>([]);
+//   const [user, setUser] = useState<User[]>([]);
+
+//   const client = new InstillClient(
+//     "https://api.instill.tech",
+//     "v1alpha",
+//     "" // console API token
+//   );
+
+//   useEffect(() => {
+//     client.Auth.getUserQuery()
+//       .then((data: any) => {
+//         console.log("data", data);
+//         setUser(data);
+//       })
+//       .catch((error: any) => {
+//         console.log("error", error);
+//       });
+
+//     client.Pipeline.listPipelinesQuery({
+//       pageSize: null,
+//       nextPageToken: null,
+//     })
+//       .then((data: any) => {
+//         console.log("data", data);
+//         setPipelines(data);
+//       })
+//       .catch((error: any) => {
+//         console.log("error", error);
+//       });
+//   }, []);
+
+//   return (
+//     <>
+//       <h1>User Data</h1>
+//       <pre style={{ backgroundColor: "white" }}>
+//         {JSON.stringify(user, null, 4)}
+//       </pre>
+
+//       <h1>Pipelines List</h1>
+//       <pre style={{ backgroundColor: "white" }}>
+//         {JSON.stringify(pipelines, null, 4)}
+//       </pre>
+//     </>
+//   );
+// }
