+++ examples/node-app/package.json
@@ -4,8 +4,7 @@
   "description": "",
   "main": "index.js",
   "dependencies": {
-    "@instill-ai/typescript-sdk": "^0.0.7",
-    "dotenv": "^16.3.1",
+    "@instill-ai/typescript-sdk": "0.0.8-rc.12",
     "express": "^4.18.2"
   },
   "scripts": {
