+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@instill-ai/typescript-sdk",
-  "version": "0.0.8-rc.12",
+  "version": "0.0.8-rc.13",
   "description": "Instill AI's Typescript SDK",
   "repository": "https://github.com/instill-ai/typescript-sdk.git",
   "bugs": "https://github.com/instill-ai/community/issues",
