+++ src/connector/types.ts
@@ -1,4 +1,5 @@
 /* eslint-disable  @typescript-eslint/no-explicit-any */
+import { AirbyteFieldValues } from "@instill-ai/toolkit";
 import { Spec } from "../types";
 
 export type ConnectorResourceState =
@@ -73,3 +74,72 @@ export type ConnectorResourcesWatchState = Record<
 export type ConnectorResourceWithWatchState = {
   watchState: ConnectorResourceState;
 } & ConnectorResourceWithDefinition;
+
+export type TestUserConnectorResourceConnectionResponse = {
+  state: ConnectorResourceState;
+};
+
+export type ConnectUserConnectorResourceResponse = {
+  connector_resource: ConnectorResource;
+};
+
+export type DisconnectUserConnectorResourceResponse = {
+  connector_resource: ConnectorResource;
+};
+
+export type CreateUserConnectorResourcePayload = {
+  id: string;
+  connector_definition_name: string;
+  description?: string;
+  configuration: Record<string, any> | Record<string, never>;
+};
+
+export type CreateUserConnectorResourceResponse = {
+  connector_resource: ConnectorResource;
+};
+
+export type UpdateUserConnectorResourcePayload = {
+  connectorResourceName: string;
+  description?: string;
+  configuration: /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
+  Record<string, any> | AirbyteFieldValues | Record<string, never>;
+};
+
+export type UpdateUserConnectorResourceResponse = {
+  connector_resource: ConnectorResource;
+};
+
+export type RenameUserConnectorResourcePayload = {
+  name: string;
+  new_connector_id: string;
+};
+
+export type RenameUserConnectorResourceResponse = {
+  connector_resource: ConnectorResource;
+};
+
+export type ListConnectorResourcesResponse = {
+  connector_resources: ConnectorResourceWithDefinition[];
+  next_page_token: string;
+  total_size: number;
+};
+
+export type ListUserConnectorResourcesResponse = {
+  connector_resources: ConnectorResourceWithDefinition[];
+  next_page_token: string;
+  total_size: number;
+};
+
+export type ListConnectorDefinitionsResponse = {
+  connector_definitions: ConnectorDefinition[];
+  next_page_token: string;
+  total_size: number;
+};
+
+export type GetConnectorDefinitionResponse = {
+  connector_definition: ConnectorDefinition;
+};
+
+export type GetUserConnectorResourceResponse = {
+  connector_resource: ConnectorResourceWithDefinition;
+};
