+++ src/metric/pipeline/index.ts
@@ -1,2 +1,1 @@
-export * from "./queries";
 export * from "./types";
