--- src/model/actions.ts
@@ -1,48 +1,0 @@
-import { createClient } from "../helper";
-import { Nullable } from "../types";
-
-export type DeployUserModelResponse = {
-  model_id: string;
-};
-
-export async function deployUserModelAction({
-  modelName,
-  accessToken,
-}: {
-  modelName: string;
-  accessToken: Nullable<string>;
-}) {
-  try {
-    const client = createClient(accessToken, "model");
-
-    const { data } = await client.post<DeployUserModelResponse>(
-      `/${modelName}/deploy`
-    );
-    return Promise.resolve(data.model_id);
-  } catch (err) {
-    return Promise.reject(err);
-  }
-}
-
-export type UndeployUserModelResponse = {
-  model_id: string;
-};
-
-export async function undeployUserModeleAction({
-  modelName,
-  accessToken,
-}: {
-  modelName: string;
-  accessToken: Nullable<string>;
-}) {
-  try {
-    const client = createClient(accessToken, "model");
-
-    const { data } = await client.post<UndeployUserModelResponse>(
-      `/${modelName}/undeploy`
-    );
-    return Promise.resolve(data.model_id);
-  } catch (err) {
-    return Promise.reject(err);
-  }
-}
