+++ src/model/types.ts
@@ -1,7 +1,8 @@
 /* eslint-disable  @typescript-eslint/no-explicit-any */
 
+import { Operation } from "../operation";
 import { Pipeline } from "../pipeline";
-import { Visibility } from "../types";
+import { Nullable, Visibility } from "../types";
 
 export type ModelReleaseStage =
   | "RELEASE_STAGE_UNSPECIFIED"
@@ -91,4 +92,113 @@ export type ModelsWatchState = Record<string, ModelWatchState>;
 
 export type ModelWithPipelines = Model & {
   pipelines: Pipeline[];
+};
+
+export type DeployUserModelResponse = {
+  model_id: string;
+};
+
+export type UndeployUserModelResponse = {
+  model_id: string;
+};
+
+export type CreateUserGithubModelConfiguration = {
+  repository: string;
+  tag: string;
+};
+
+export type CreateUserLocalModelConfiguration = {
+  content: File;
+};
+
+export type CreateUserArtivcModelConfiguration = {
+  url: string;
+  tag: string;
+  credential: Nullable<string>;
+};
+
+export type CreateUserHuggingFaceModelConfiguration = {
+  repo_id: string;
+};
+
+export type CreateUserModelPayload =
+  | CreateUserGitHubModelPayload
+  | CreateUserLocalModelPayload
+  | CreateUserHuggingFaceModelPayload
+  | CreateUserArtiVCModelPayload;
+
+export type CreateUserGitHubModelPayload = {
+  id: string;
+  model_definition: string;
+  description?: string;
+  type: "GitHub";
+  configuration: CreateUserGithubModelConfiguration;
+};
+
+export type CreateUserLocalModelPayload = {
+  id: string;
+  model_definition: string;
+  description?: string;
+  type: "Local";
+  configuration: CreateUserLocalModelConfiguration;
+};
+
+export type CreateUserHuggingFaceModelPayload = {
+  id: string;
+  model_definition: string;
+  description?: string;
+  type: "HuggingFace";
+  configuration: CreateUserHuggingFaceModelConfiguration;
+};
+
+export type CreateUserArtiVCModelPayload = {
+  id: string;
+  model_definition: string;
+  description?: string;
+  type: "ArtiVC";
+  configuration: CreateUserArtivcModelConfiguration;
+};
+
+export type CreateUserModelResponse = {
+  operation: Operation;
+};
+
+export type UpdateUserModelPayload = {
+  name: string;
+  description?: string;
+  configuration?: Record<string, any>;
+};
+
+export type UpdateUserModelResponse = {
+  model: Model;
+};
+
+export type GetModelDefinitionResponse = {
+  model_definition: ModelDefinition;
+};
+
+export type ListModelDefinitionsResponse = {
+  model_definitions: ModelDefinition[];
+  next_page_token: string;
+  total_size: number;
+};
+
+export type GetUserModelResponse = {
+  model: Model;
+};
+
+export type ListModelsResponse = {
+  models: Model[];
+  next_page_token: string;
+  total_size: number;
+};
+
+export type ListUserModelsResponse = {
+  models: Model[];
+  next_page_token: string;
+  total_size: number;
+};
+
+export type GetUserModelReadmeQueryResponse = {
+  readme: ModelReadme;
 };
