--- src/operation/queries.ts
@@ -1,54 +1,0 @@
-import { Operation } from "./types";
-import { createClient } from "../helper";
-import { Nullable } from "../types";
-
-export type GetModelOperationResponse = {
-  operation: Operation;
-};
-
-export async function getOperationQuery({
-  operationName,
-  accessToken,
-}: {
-  operationName: string;
-  accessToken: Nullable<string>;
-}) {
-  try {
-    const client = createClient(accessToken, "model");
-
-    const { data } = await client.get<GetModelOperationResponse>(
-      `/${operationName}`
-    );
-    return Promise.resolve(data.operation);
-  } catch (err) {
-    return Promise.reject(err);
-  }
-}
-
-export async function checkUntilOperationIsDoen({
-  operationName,
-  accessToken,
-}: {
-  operationName: string;
-  accessToken: Nullable<string>;
-}): Promise<boolean> {
-  try {
-    const operation = await getOperationQuery({
-      operationName,
-      accessToken,
-    });
-
-    if (operation.done) {
-      return Promise.resolve(true);
-    } else {
-      await new Promise((resolve) => setTimeout(resolve, 1000));
-      const result = await checkUntilOperationIsDoen({
-        operationName,
-        accessToken,
-      });
-      return Promise.resolve(result);
-    }
-  } catch (err) {
-    return Promise.reject(err);
-  }
-}
